Refactored main robot code to use 1816 name space.
The main libs were left pretty much intact and most of the main rebod code remained the same just change the package names and talon ids.

Part of this refactor included enhancing the TalonSRXFactory.  I created a concept of a GhostTalonSRX.java this is a dummy TalonSRX that has the same interface as a TalonSRX but does pretty much nothing.  This allows the code to run as as without having the exact same robot.  You just put a -1 for any Talon ids for Talons that you do not have.  See Constants.java

The original repo was converted to use the 2019 control system